+++ PiezasProveedores/src/main/java/T26/PiezasProveedores/dao/ISuministraDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import T26.PiezasProveedores.dto.*;
 
-public interface ISuministraDAO extends JpaRepository<Suministra, Long> {
+public interface ISuministraDAO extends JpaRepository<Suministra, Integer> {
 
 }
